+++ consensus/qbft/src/main/java/org/hyperledger/besu/consensus/qbft/validation/ProposalPayloadValidator.java
@@ -112,9 +112,6 @@ public class ProposalPayloadValidator {
     return true;
   }
 
-  /*
-    Validate CMS in block header (only for PKI-mode)
-  */
   private boolean validateCms(
       final Block block,
       final BftBlockInterface bftBlockInterface,